Implementing suggestions from Pull-Request
