+++ Cargo.toml
@@ -25,8 +25,7 @@ rustdoc-args = ["--cfg", "feature=\"docs\""]
 [features]
 default = ["hyper-server"]
 hyper-server = ["http-service-hyper"]
-docs = ["unstable", "cookies"]
-cookies = ["cookie"]
+docs = ["unstable"]
 unstable = []
 
 [dependencies]
@@ -42,7 +41,7 @@ serde_qs = "0.5.0"
 async-std = { version = "1.0.1", features = ["unstable"] }
 pin-project-lite = "0.1.0"
 mime = "0.3.14"
-cookie = { version="0.12.0", features = ["percent-encode"], optional = true }
+cookie = { version="0.12.0", features = ["percent-encode"]}
 
 [dev-dependencies]
 async-std = { version = "1.0.1", features = ["unstable", "attributes"] }
