+++ src/middleware/cookies.rs
@@ -5,14 +5,10 @@ use crate::{
 use cookie::{Cookie, CookieJar, ParseError};
 use futures::future::BoxFuture;
 
-use crate::{error::ResultExt, error::StringError, Error};
-use http::{status::StatusCode, HeaderMap};
+use http::HeaderMap;
 use std::sync::{Arc, RwLock};
 
-const MIDDLEWARE_MISSING_MSG: &str =
-    "CookiesMiddleware must be used to populate request and response cookies";
-
-/// A simple requests logger
+/// A middleware for making cookie data available in requests.
 ///
 /// # Examples
 ///
@@ -69,7 +65,7 @@ pub(crate) struct CookieData {
 }
 
 impl CookieData {
-    pub fn from_headers(headers: &HeaderMap) -> Self {
+    pub(crate) fn from_headers(headers: &HeaderMap) -> Self {
         CookieData {
             content: Arc::new(RwLock::new(
                 headers
@@ -90,185 +86,4 @@ fn parse_from_header(s: &str) -> Result<CookieJar, ParseError> {
     })?;
 
     Ok(jar)
-}
-
-/// An extension to `Context` that provides cached access to cookies
-pub trait RequestExt {
-    /// returns a `Cookie` by name of the cookie
-    fn get_cookie(&self, name: &str) -> Result<Option<Cookie<'static>>, Error>;
-
-    /// Add cookie to the cookie jar
-    fn set_cookie(&mut self, cookie: Cookie<'static>) -> Result<(), Error>;
-
-    /// Removes the cookie. This instructs the `CookiesMiddleware` to send a cookie with empty value
-    /// in the response.
-    fn remove_cookie(&mut self, cookie: Cookie<'static>) -> Result<(), Error>;
-}
-
-impl<State> RequestExt for Request<State> {
-    fn get_cookie(&self, name: &str) -> Result<Option<Cookie<'static>>, Error> {
-        let cookie_data = self
-            .local::<CookieData>()
-            .ok_or_else(|| StringError(MIDDLEWARE_MISSING_MSG.to_owned()))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-
-        let locked_jar = cookie_data
-            .content
-            .read()
-            .map_err(|e| StringError(format!("Failed to get read lock: {}", e)))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-
-        Ok(locked_jar.get(name).cloned())
-    }
-
-    fn set_cookie(&mut self, cookie: Cookie<'static>) -> Result<(), Error> {
-        let cookie_data = self
-            .local::<CookieData>()
-            .ok_or_else(|| StringError(MIDDLEWARE_MISSING_MSG.to_owned()))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-
-        let mut locked_jar = cookie_data
-            .content
-            .write()
-            .map_err(|e| StringError(format!("Failed to get read lock: {}", e)))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-        locked_jar.add(cookie);
-        Ok(())
-    }
-
-    fn remove_cookie(&mut self, cookie: Cookie<'static>) -> Result<(), Error> {
-        let cookie_data = self
-            .local::<CookieData>()
-            .ok_or_else(|| StringError(MIDDLEWARE_MISSING_MSG.to_owned()))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-
-        let mut locked_jar = cookie_data
-            .content
-            .write()
-            .map_err(|e| StringError(format!("Failed to get read lock: {}", e)))
-            .with_err_status(StatusCode::INTERNAL_SERVER_ERROR)?;
-        locked_jar.remove(cookie);
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use cookie::Cookie;
-    use futures::executor::block_on;
-    use futures::AsyncReadExt;
-    use http_service::Body;
-    use http_service_mock::make_server;
-
-    static COOKIE_NAME: &str = "testCookie";
-
-    async fn retrieve_cookie(cx: Request<()>) -> String {
-        cx.get_cookie(COOKIE_NAME)
-            .unwrap()
-            .unwrap()
-            .value()
-            .to_string()
-    }
-
-    async fn set_cookie(mut cx: Request<()>) -> Response {
-        cx.set_cookie(Cookie::new(COOKIE_NAME, "NewCookieValue"))
-            .unwrap();
-        Response::new(200)
-    }
-
-    async fn remove_cookie(mut cx: Request<()>) -> Response {
-        cx.remove_cookie(Cookie::named(COOKIE_NAME)).unwrap();
-        Response::new(200)
-    }
-
-    async fn set_multiple_cookie(mut cx: Request<()>) -> Response {
-        cx.set_cookie(Cookie::new("C1", "V1")).unwrap();
-        cx.set_cookie(Cookie::new("C2", "V2")).unwrap();
-        Response::new(200)
-    }
-
-    fn app() -> crate::Server<()> {
-        let mut app = crate::new();
-        app.middleware(CookiesMiddleware::new());
-
-        app.at("/get").get(retrieve_cookie);
-        app.at("/set").get(set_cookie);
-        app.at("/remove").get(remove_cookie);
-        app.at("/multi").get(set_multiple_cookie);
-        app
-    }
-
-    fn make_request(endpoint: &str) -> http::response::Response<http_service::Body> {
-        let app = app();
-        let mut server = make_server(app.into_http_service()).unwrap();
-        let req = http::Request::get(endpoint)
-            .header(http::header::COOKIE, "testCookie=RequestCookieValue")
-            .body(Body::empty())
-            .unwrap();
-        server.simulate(req).unwrap()
-    }
-
-    #[test]
-    fn successfully_retrieve_request_cookie() {
-        let mut res = make_request("/get");
-        assert_eq!(res.status(), 200);
-
-        let body = block_on(async move {
-            let mut buffer = Vec::new();
-            res.body_mut().read_to_end(&mut buffer).await.unwrap();
-            buffer
-        });
-
-        assert_eq!(&*body, &*b"RequestCookieValue");
-    }
-
-    #[test]
-    fn successfully_set_cookie() {
-        let res = make_request("/set");
-        assert_eq!(res.status(), 200);
-        let test_cookie_header = res.headers().get(http::header::SET_COOKIE).unwrap();
-        assert_eq!(
-            test_cookie_header.to_str().unwrap(),
-            "testCookie=NewCookieValue"
-        );
-    }
-
-    #[test]
-    fn successfully_remove_cookie() {
-        let res = make_request("/remove");
-        assert_eq!(res.status(), 200);
-        let test_cookie_header = res.headers().get(http::header::SET_COOKIE).unwrap();
-        assert!(test_cookie_header
-            .to_str()
-            .unwrap()
-            .starts_with("testCookie=;"));
-        let cookie = Cookie::parse_encoded(test_cookie_header.to_str().unwrap()).unwrap();
-        assert_eq!(cookie.name(), COOKIE_NAME);
-        assert_eq!(cookie.value(), "");
-        assert_eq!(cookie.http_only(), None);
-        assert_eq!(cookie.max_age().unwrap().num_nanoseconds(), Some(0));
-    }
-
-    #[test]
-    fn successfully_set_multiple_cookies() {
-        let res = make_request("/multi");
-        assert_eq!(res.status(), 200);
-        let cookie_header = res.headers().get_all(http::header::SET_COOKIE);
-        let mut iter = cookie_header.iter();
-
-        let cookie1 = iter.next().unwrap();
-        let cookie2 = iter.next().unwrap();
-
-        //Headers can be out of order
-        if cookie1.to_str().unwrap().starts_with("C1") {
-            assert_eq!(cookie1, "C1=V1");
-            assert_eq!(cookie2, "C2=V2");
-        } else {
-            assert_eq!(cookie2, "C1=V1");
-            assert_eq!(cookie1, "C2=V2");
-        }
-
-        assert!(iter.next().is_none());
-    }
 }
